Подключение MapStruct. доработка конфигурации maven. Подготовка мапперов.
