+++ src/main/java/com/example/coffemachine/model/dto/IngredientDTO.java
@@ -1,17 +1,17 @@
-package com.example.coffemachine.models.dto;
+package com.example.coffemachine.model.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Set;
-
 @Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class DrinkDTO {
+public class IngredientDTO {
 
     private Long id;
     private String name;
-    private Set<IngredientDTO> ingredients;
 }
+
